Revert "Feature/authentication"

+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/User.kt
@@ -14,7 +14,8 @@ data class User(
     val workoutDays: List<WorkoutDay>
 )
 enum class Gender {
-    MALE, FEMALE}
+    MALE, FEMALE
+}
 
 enum class MeasurementUnit {
     METRIC, IMPERIAL
@@ -25,7 +26,7 @@ enum class FitnessGoal {
 }
 
 data class Tool(
-    val id: Long,
+    val id: Int,
     val name: String
 )
